TravelAgencySystem/src/com/travelagency/Passenger.java
package com.travelagency;

public class Passenger {
    private String name;
    private int passengerNumber;
    private double balance;

    public Passenger(String name, int passengerNumber) {
        this.name = name;
        this.passengerNumber = passengerNumber;
        this.balance = 0;
    }

    public String getName() {
        return name;
    }

    public int getPassengerNumber() {
        return passengerNumber;
    }

    public double getBalance() {
        return balance;
    }

    public void setBalance(double balance) {
        this.balance = balance;
    }
}
